Added Stuff
You should now recieve all the data you need to save.  Take a look at
the save function in the controller. It gives you the data as a json.

Add course and click "save" and the code will run. It looks a little
messy atm, and i am not sure why that is.

All that is left for the save part is for you to finish up the
controller stuff.

I also need the information that was previously saved to load by default
So youll have to let me know when this gets done. Feel free to message
me with questions