Added stuff
Adding page centering so when you zoom way out it will center it. (Give
it a try)
Adding hovering function for professors. (still waiting for these to get
added to database)
Added a button to go home. I experimented with there styleing, but i
dont think it is gonna work so i am adding the navigation in.
Change javascript so that "save" works when directed off a controller.
This is how it will function in the actual app. So just keep in mind if
you load AdvisingView as the default controller, "save" will fail.