Read Description
Most of the changes here are for navigation.
You now should log in as a student or advisor for testing from now on.
1. Changes routes to default to 'Login'
2. Uncommented session and commented out the static student id in the
advisingform controller.
3.Add a function in our controller for loadAllStudents.
It is for the advisor view that loads all the students for the professor
to select.
4. Add view for this as well. Its templete is almost complete. Still
have to add the students which should get passed in with SID's from our
controller.
5. Added link in student main page to get to advising form