Read Description
Advisor should now be able to load a students advising_form. (there are
some errors, but all my stuff is correct.) Pls log in as gourd, select
as student and troubleshoot these errors.

I also modified links so that php echos it out correctly everytime.

I moved all of the functions for the advisior stuff to the mainpage.